Renames and minor changes
