+++ src/components/button-scroll-top.vue
@@ -19,7 +19,6 @@ export default {
         top:0,
         behavior: 'smooth'
       });
-
     }
   }
 }
